Simplify converting OssShipperConfig to JSON object (#6)
Currently, when converting the object storageDetail to a JSON object, we cast it to the corresponding class type based on it's storage format. However, what ever the actual class it is, we only need to call it's ToJsonObject method which is a abstract method in the base class. So we can simplify it as JSONObject obj = storageDetail.ToJsonObject();